Move `data.calculate|filter` to `transform.calculate|filter`  so we have clean separate between `data` source and data `transform`.
