+++ examples/transparent_layer_bar.json
@@ -1,6 +1,7 @@
 {
   "description": "A bar chart showing the US population distribution of age groups and gender in 2000.",
-  "data": { "url": "data/population.json",
+  "data": { "url": "data/population.json"},
+  "transform": {
     "filter": "datum.year == 2000",
     "calculate": [{"field": "gender", "expr": "datum.sex == 2 ? \"Female\" : \"Male\""}]
   },
