+++ src/schema/schema.ts
@@ -6,11 +6,13 @@ import {config, Config} from './config.schema';
 import {data, Data} from './data.schema';
 import {encoding, Encoding} from './encoding.schema';
 import {Mark} from '../mark';
+import {transform, Transform} from './transform.schema';
 
 export interface Spec {
   name?: string;
   description?: string;
   data?: Data;
+  transform?: Transform;
   mark?: Mark;
   encoding?: Encoding;
   config?: Config;
@@ -36,6 +38,7 @@ export var schema = {
       type: 'string'
     },
     data: data,
+    transform: transform,
     mark: mark,
     encoding: encoding,
     config: config
