+++ tsconfig.json
@@ -72,6 +72,7 @@
     "src/schema/schemagen.ts",
     "src/schema/schemautil.ts",
     "src/schema/sort.schema.ts",
+    "src/schema/transform.schema.ts",
     "src/shorthand.ts",
     "src/spec.ts",
     "src/timeunit.ts",
